fixed a warning

+++ reporter.hpp
@@ -386,7 +386,7 @@ namespace reporter {
         static std::string repeat(std::string s, size_t n) 
         { 
             std::string s1 = s; 
-            for (int i=1; i<n;i++)  s += s1;
+            for (size_t i=1; i<n;i++)  s += s1;
             return s; 
         }
